add user auth type to user model
